requirements.txt
langchain 
openai 
tiktoken 
chromadb 
pypdf 
pyautogen
huggingface_hub
sentence-transformers